+++ express-zod-api/src/routing-walker.ts
@@ -38,11 +38,10 @@ const trimPath = (path: string) =>
   path.trim().split("/").filter(Boolean).join("/");
 
 const processEntries = (
-  { methodLikeRouteBehavior = "method" }: CommonConfig,
+  { recognizeMethodDependentRoutes: preferMethod = true }: CommonConfig,
   subject: Routing,
   parent?: string,
 ) => {
-  const preferMethod = methodLikeRouteBehavior === "method";
   return Object.entries(subject).map<[string, Routing[string], Method?]>(
     ([_key, item]) => {
       const [segment, method] =