+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### v27.2.0
 
 - Supporting TypeScript 6:
-  - Supported `typescript` versions: `^5.1.3 || 6.0.2`.
+  - Supported `typescript` versions: `^5.1.3 || ^6.0.2`.
 
 ### v27.1.0